feat: group_schedule_info entity 생성
