+++ src/main/java/com/pro/domain/group_schedule_info/domain/Group_schedule_info.java
@@ -1,4 +1,4 @@
-package com.pro.domain.comment;
+package com.pro.domain.group_schedule_info.domain;
 
 import com.pro.domain.common.BaseEntity;
 import jakarta.persistence.*;
@@ -10,17 +10,13 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends BaseEntity {
+@Table(name = "t_mn_group_schedule_info")
+public class Group_schedule_info extends BaseEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "comment_id")
+  @Column(name = "group_schedule_id")
   private Long id;
 
-  @NotBlank(message = "일정 아이디 값은 필수입니다.")
-  private Long schedule_id;
-
-  @NotBlank
-  @Column(columnDefinition = "TEXT")//65535자의 고정된 최대 크기
-  private String content; //댓글내용
-
+  @NotBlank(message = "그룹 아이디는 필수값 입니다.")
+  private Long group_id;
 }
