+++ src/main/java/com/pro/domain/schedule/domain/Schedule.java
@@ -12,6 +12,7 @@ import java.time.LocalDateTime;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "t_mn_schedule")
 public class Schedule extends BaseEntity {
 
   @Id
